refactor(agent-core): merge pr-risk-classifier into change-type-classifier
Move isLowRiskPR and reviewersForDiff into change-type-classifier.ts,
deriving isLowRiskPR from classifyChanges().type membership in the
LOW_RISK_TYPES set. Eliminates duplicate file-predicate tables.

Closes #1953
