+++ packages/agent-core/src/change-type-classifier.ts
@@ -104,7 +104,33 @@ const CLASSIFICATION_RULES: readonly ClassificationRule[] = [
   },
 ] as const;
 
+// ── Constants ───────────────────────────────────────────────────────
+
+/** Change types that are safe to auto-merge once CI passes. */
+const LOW_RISK_TYPES: ReadonlySet<ChangeType> = new Set([
+  "dependency",
+  "docs",
+  "config",
+  "test",
+  "infrastructure",
+]);
+
 // ── Core functions ──────────────────────────────────────────────────
+
+/**
+ * Returns `true` when every file in `files` classifies as a low-risk change
+ * type (dependency, docs, config, test, or infrastructure) and the PR can be
+ * auto-merged immediately once CI passes.
+ *
+ * Returns `false` for an empty file list, mixed-type changes, or any list
+ * containing frontend or backend source files.
+ */
+export function isLowRiskPR(files: readonly string[]): boolean {
+  if (files.length === 0) {
+    return false;
+  }
+  return LOW_RISK_TYPES.has(classifyChanges(files).type);
+}
 
 /**
  * Classifies a list of changed files into a single change type and determines
@@ -195,4 +221,83 @@ function buildReason(
   }
 
   return `${base} — skipping: ${skipPhases.join(", ")}`;
+}
+
+// ── Diff-matched specialized reviewers ───────────────────────────────────────
+
+/**
+ * Maps a changed-file predicate to the specialized review agent that should
+ * inspect a PR touching those files. Each reviewer catches a class of semantic
+ * regression that CI (lint/typecheck/test) cannot. Order here is the order
+ * `reviewersForDiff` returns them in — keep it intentional.
+ */
+const DIFF_REVIEWERS: ReadonlyArray<{ name: string; matches: (file: string) => boolean }> = [
+  {
+    // Prisma schema or migration SQL — destructive/semantic migration review.
+    name: "migration-reviewer",
+    matches: (f) => /\.prisma$/.test(f) || /(?:^|\/)migrations?\/.*\.sql$/i.test(f),
+  },
+  {
+    // Server code and the edge router — most governed by active ADRs
+    // (RFC-7807 errors, api-versioning, auth). Excludes tests.
+    name: "adr-compliance-reviewer",
+    matches: (f) =>
+      (/^services\/[^/]+\/src\//.test(f) || /^infrastructure\/worker\/src\//.test(f)) &&
+      !/\.(test|spec)\.[tj]sx?$/.test(f),
+  },
+  {
+    // Rialto component or component-test changes — prop-drift between a
+    // component's *Props interface and its tests.
+    name: "rialto-prop-drift-detector",
+    matches: (f) =>
+      f.startsWith("packages/rialto/src/components/") ||
+      f.startsWith("packages/rialto/src/test/") ||
+      /^packages\/rialto\/src\/.*\.(test|spec)\.tsx$/.test(f),
+  },
+  {
+    // Dependency manifests — version-bump safety across the monorepo.
+    name: "dependency-update-reviewer",
+    matches: (f) =>
+      f === "package.json" ||
+      f.endsWith("/package.json") ||
+      f === "pnpm-lock.yaml" ||
+      f === "pnpm-workspace.yaml",
+  },
+  {
+    // Payment, webhook, deposit, and Stripe-touching files — catches missing
+    // webhook signature verification, dropped idempotency keys, float/integer
+    // money handling, and broken refund/deposit/charge state transitions that
+    // CI regex and schema checks cannot surface.
+    name: "stripe-flow-reviewer",
+    matches: (f) =>
+      /(?:^|\/)(?:payment|payments|webhook|webhooks|deposit|deposits|stripe|charge|charges|refund|refunds)[^/]*\.[tj]sx?$/.test(
+        f
+      ),
+  },
+  {
+    // Generated artifacts (llms.txt context bundles, generated zod schemas,
+    // dependency graph) and the pack generator that emits them. These drift
+    // non-deterministically across platforms (locale-sensitive sorts) or go
+    // stale relative to their source — a class of Integrity failure CI only
+    // catches after push. See the byte-order-comparator fix for the root cause.
+    name: "generated-artifact-determinism-reviewer",
+    matches: (f) =>
+      f === "tools/cli/src/commands/pack.ts" ||
+      f === "llms.txt" ||
+      f === "llms-full.txt" ||
+      f.endsWith("/llms.txt") ||
+      f.endsWith("/llms-full.txt") ||
+      f.endsWith("/generated-schemas.ts") ||
+      f === "infrastructure/worker/dep-graph.json" ||
+      f === "docs/architecture/dependency-graph.md",
+  },
+];
+
+/**
+ * Returns the specialized review agents whose trigger matches at least one
+ * changed file, de-duplicated and in a stable order. Most PRs match zero or
+ * one reviewer. Used by implement-queue to gate a PR before the merge train.
+ */
+export function reviewersForDiff(files: readonly string[]): string[] {
+  return DIFF_REVIEWERS.filter((r) => files.some((f) => r.matches(f))).map((r) => r.name);
 }
